Working test with simulations
